Update Laboratorio.cpp Fix warning

+++ practica1/src/Laboratorio.cpp
@@ -29,7 +29,7 @@ Laboratorio::Laboratorio(int id, const std::string &nombre_lab, const std::strin
 }
 
 bool Laboratorio::operator==(Laboratorio lab) const{
-    if(m_id == lab.getId()){
+    if(static_cast<int>(m_id) == lab.getId()){
         return true;
     }
     return false;
